# app/docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This swagger describes every endpoint of this project.",
        "title": "Augustin Swagger",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "GNU Affero General Public License",
            "url": "https://www.gnu.org/licenses/agpl-3.0.txt"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "paths": {
...
        "/map/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Get longitudes and latitudes of all vendors for online map",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Map"
                ],
                "summary": "Get longitudes and latitudes of all vendors for online map",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.LocationData"
                            }
                        }
                    }
                }
            }
        },
...
        "/vendors/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "List Vendors",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/database.Vendor"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Create Vendor",
                "parameters": [
                    {
                        "description": "Vendor Representation",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Vendor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/vendors/check/{licenseID}/": {
            "get": {
                "description": "Check if license id exists, return first name of vendor if it does",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Check for license id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "License ID",
                        "name": "licenseID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/vendors/me/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Get Vendor overview",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.VendorOverview"
                        }
                    }
                }
            }
        },
        "/vendors/{id}/": {
            "get": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Get Vendor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "description": "Warning: Unfilled fields will be set to default values",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Update Vendor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Vendor Representation",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.Vendor"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "KeycloakAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vendors"
                ],
                "summary": "Delete Vendor",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vendor ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.LocationData": {
            "type": "object",
            "properties": {
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latitude": {
                    "type": "number"
                },
                "licenseID": {
                    "type": "string"
                },
                "longitude": {
                    "type": "number"
                }
            }
        },
...
        "database.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "authorizedBy": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "isPayoutFor": {
                    "description": "Connected payout payment",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Payment"
                    }
                },
                "isSale": {
                    "type": "boolean"
                },
                "item": {
                    "type": "integer"
                },
                "order": {
                    "type": "integer"
                },
                "orderEntry": {
                    "type": "integer"
                },
                "payout": {
                    "description": "Connected payout payment",
                    "type": "integer"
                },
                "price": {
                    "description": "Price at time of purchase in cents",
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "receiver": {
                    "type": "integer"
                },
                "receiverName": {
                    "description": "JOIN from Receiver Account",
                    "allOf": [
                        {
                            "$ref": "#/definitions/null.String"
                        }
                    ]
                },
                "sender": {
                    "type": "integer"
                },
                "senderName": {
                    "description": "JOIN from Sender Account",
                    "allOf": [
                        {
                            "$ref": "#/definitions/null.String"
                        }
                    ]
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "database.Vendor": {
            "type": "object",
            "properties": {
                "accountProofUrl": {
                    "$ref": "#/definitions/null.String"
                },
                "balance": {
                    "description": "This is joined in from the account",
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Comment"
                    }
                },
                "debt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "hasBankAccount": {
                    "type": "boolean"
                },
                "hasSmartphone": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "isDeleted": {
                    "type": "boolean"
                },
                "isDisabled": {
                    "type": "boolean"
                },
                "keycloakID": {
                    "type": "string"
                },
                "language": {
                    "type": "string"
                },
                "lastName": {
                    "type": "string"
                },
                "lastPayout": {
                    "type": "string",
                    "format": "date-time"
                },
                "licenseID": {
                    "$ref": "#/definitions/null.String"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Location"
                    }
                },
                "onlineMap": {
                    "type": "boolean"
                },
                "registrationDate": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                },
                "urlID": {
                    "description": "This is used for the QR code",
                    "type": "string"
                },
                "vendorSince": {
                    "type": "string"
                }
            }
        },
        "ent.Comment": {
            "type": "object",
            "properties": {
                "comment": {
                    "description": "Comment holds the value of the \"comment\" field.",
                    "type": "string"
                },
                "created_at": {
                    "description": "CreatedAt holds the value of the \"created_at\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CommentQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.CommentEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "resolved_at": {
                    "description": "ResolvedAt holds the value of the \"resolved_at\" field.",
                    "type": "string"
                },
                "warning": {
                    "description": "Warning holds the value of the \"warning\" field.",
                    "type": "boolean"
                }
            }
        },
        "ent.CommentEdges": {
            "type": "object",
            "properties": {
                "vendor": {
                    "description": "Vendor holds the value of the vendor edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Vendor"
                        }
                    ]
                }
            }
        },
        "ent.Location": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address holds the value of the \"address\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the LocationQuery when eager-loading is set.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.LocationEdges"
                        }
                    ]
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "latitude": {
                    "description": "Latitude holds the value of the \"latitude\" field.",
                    "type": "number"
                },
                "longitude": {
                    "description": "Longitude holds the value of the \"longitude\" field.",
                    "type": "number"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "working_time": {
                    "description": "WorkingTime holds the value of the \"working_time\" field.",
                    "type": "string"
                },
                "zip": {
                    "description": "Zip holds the value of the \"zip\" field.",
                    "type": "string"
                }
            }
        },
        "ent.LocationEdges": {
            "type": "object",
            "properties": {
                "vendor": {
                    "description": "Vendor holds the value of the vendor edge.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/ent.Vendor"
                        }
                    ]
                }
            }
        },
...
        "handlers.VendorOverview": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "integer"
                },
                "comments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Comment"
                    }
                },
                "email": {
                    "type": "string"
                },
                "firstName": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastName": {
                    "type": "string"
                },
                "lastPayout": {
                    "type": "string",
                    "format": "date-time"
                },
                "licenseID": {
                    "type": "string"
                },
                "locations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Location"
                    }
                },
                "openPayments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/database.Payment"
                    }
                },
                "plz": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                },
                "urlID": {
                    "type": "string"
                }
            }
        },
...
        "null.String": {
            "type": "object",
            "properties": {
                "string": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if String is not NULL",
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "KeycloakAuth": {
            "description": "\u003cb\u003ehow to generate an api key\u003c/b\u003e \u003cbr/\u003e\u003cbr/\u003e\u003ccode\u003ecurl -d 'client_id=frontend' -d 'scope=openid' -d 'username=test_superuser' -d 'password=Test123!' -d 'grant_type=password' 'http://keycloak:8080/realms/augustin/protocol/openid-connect/token' |     python3 -m json.tool | grep access_token\u003c/code\u003e\u003cbr/\u003e\u003cbr/\u003e\u003cbr/\u003eInsert the output into the field below the value of the access_token field.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}